db: Rename JSONType to JSON, remove AsJSONNull
